swap in new navigation
