+++ navigation/index-prev.tsx
@@ -1,35 +1,45 @@
-import { createNativeStackNavigator } from '@react-navigation/native-stack'
+/**
+ * If you are not familiar with React Navigation, refer to the "Fundamentals" guide:
+ * https://reactnavigation.org/docs/getting-started
+ *
+ */
+import { FontAwesome } from '@expo/vector-icons'
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs'
+import { NavigationContainer, DefaultTheme, DarkTheme } from '@react-navigation/native'
+import { createNativeStackNavigator } from '@react-navigation/native-stack'
 import * as React from 'react'
-import { NavigationContainer } from '@react-navigation/native'
+import { ColorSchemeName, Pressable } from 'react-native'
 
+import Colors from '../constants/Colors'
+import useColorScheme from '../hooks/useColorScheme'
 import ChatRoomScreen from '../screens/Chatroom'
-import ChatsScreen from '../screens/Chats'
-import LoginScreen from '../screens/LoginScreen'
+import { ChatsScreen } from '../screens/Chats'
 import ModalScreen from '../screens/ModalScreen'
 import NotFoundScreen from '../screens/NotFoundScreen'
 import TabOneScreen from '../screens/TabOneScreen'
 import TabTwoScreen from '../screens/TabTwoScreen'
 import { RootStackParamList, RootTabParamList, RootTabScreenProps } from '../types'
 import LinkingConfiguration from './LinkingConfiguration'
-import { Pressable } from 'react-native'
-import { FontAwesome } from '@expo/vector-icons'
-import TabBarIcon from '../components/TabBarIcon'
 
-export default function Navigation() {
+export default function Navigation({ colorScheme }: { colorScheme: ColorSchemeName }) {
   return (
-    <NavigationContainer linking={LinkingConfiguration}>
+    <NavigationContainer
+      linking={LinkingConfiguration}
+      theme={colorScheme === 'dark' ? DarkTheme : DefaultTheme}>
       <RootNavigator />
     </NavigationContainer>
   )
 }
 
+/**
+ * A root stack navigator is often used for displaying modals on top of all other content.
+ * https://reactnavigation.org/docs/modal
+ */
 const Stack = createNativeStackNavigator<RootStackParamList>()
 
 function RootNavigator() {
   return (
     <Stack.Navigator>
-      <Stack.Screen name='Auth' component={AuthNavigator} options={{ headerShown: false }} />
       <Stack.Screen name='Root' component={BottomTabNavigator} options={{ headerShown: false }} />
       <Stack.Screen name='NotFound' component={NotFoundScreen} options={{ title: 'Oops!' }} />
       <Stack.Group screenOptions={{ presentation: 'modal' }}>
@@ -39,9 +49,15 @@ function RootNavigator() {
   )
 }
 
+/**
+ * A bottom tab navigator displays tab buttons on the bottom of the display to switch screens.
+ * https://reactnavigation.org/docs/bottom-tab-navigator
+ */
 const BottomTab = createBottomTabNavigator<RootTabParamList>()
 
 function BottomTabNavigator() {
+  const colorScheme = useColorScheme()
+
   return (
     <BottomTab.Navigator
       initialRouteName='Chats'
@@ -78,7 +94,12 @@ function BottomTabNavigator() {
               style={({ pressed }) => ({
                 opacity: pressed ? 0.5 : 1,
               })}>
-              <FontAwesome name='info-circle' size={25} color='white' style={{ marginRight: 15 }} />
+              <FontAwesome
+                name='info-circle'
+                size={25}
+                color={Colors[colorScheme].text}
+                style={{ marginRight: 15 }}
+              />
             </Pressable>
           ),
         })}
@@ -95,12 +116,12 @@ function BottomTabNavigator() {
   )
 }
 
-const AuthStack = createNativeStackNavigator<RootStackParamList>()
-
-function AuthNavigator() {
-  return (
-    <AuthStack.Navigator>
-      <AuthStack.Screen name='Login' component={LoginScreen} options={{ headerShown: false }} />
-    </AuthStack.Navigator>
-  )
+/**
+ * You can explore the built-in icon families and icons on the web at https://icons.expo.fyi/
+ */
+function TabBarIcon(props: {
+  name: React.ComponentProps<typeof FontAwesome>['name']
+  color: string
+}) {
+  return <FontAwesome size={30} style={{ marginBottom: -3 }} {...props} />
 }
